emulator_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-emulator_env"
version = "0.1.0"
description = "Emulator Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.0",
    # Environment-specific dependencies
    "melee",
    # .env file support (transitive via pydantic-settings, but made explicit)
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m emulator_env.server.app
server = "emulator_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["emulator_env", "emulator_env.server"]
package-dir = { "emulator_env" = ".", "emulator_env.server" = "server" }